MIPSAlign {MIPSAlign}	R Documentation

MIPSAlign trimmer and aligner

Description

Combines trimming and aligning in one comprehensive function

Usage


MIPSAlign(
  fileDir = getwd(),
  refFile,
  filter = FALSE,
  leftPattern = "",
  rightPattern = "",
  minLen = 60,
  trimTo = 40,
  aligner = c("matchAlign", "btAlign", "bt2Align"),
  numCores = 1,
  keepFiles = FALSE,
  options = ""
)

Arguments

fileDir
The file directory where all the Fastq files are located

refFile
The reference Fasta alignment file

filter
TRUE or FALSE, if you want to use the Fastq filter

leftPattern
Relevant only if filter = TRUE, the 3' adapter to trim sequences by

rightPattern
Relevant only if filter = TRUE, the 5' adapter to trim sequences by

minLen
Relevant only if filter = TRUE, all sequences below this minimum length threshold are filtered out

trimTo
Relevant only if filter = TRUE, trims all reads to desired length (determined by the length of the longest sequence)

aligner
One of "matchAlign", "btAlign", or "bt2Align"

numCores
The number of cores to run on

keepFiles
TRUE or FALSE, whether you want to keep the bowtie output files

options
Extra options for bowtie that feed into "alignmentParameter"

Value

A list of counts, unaligned counts, and possibly imperfect counts if a bowtie aligner was used

Examples


x <- MIPSAlign(refFile = fastaFileName,filter = TRUE,trimTo = 30,numCores = 6)

[Package MIPSAlign version 0.0.0.9000 Index]