bt2Align {MIPSAlign}	R Documentation

Bowtie2 aligner

Description

Generates counts using the bowtie2 algorithm

Usage


bt2Align(
  seqData = list.files(pattern = "*.fastq.gz$"),
  orfLib = list.files(pattern = "*.fa$|*.fasta$")[[1]],
  numCores = 1,
  keepFiles = FALSE,
  options = ""
)

Arguments

seqData
The list of Fastq files you want to align

orfLib
The reference Fasta alignment file

numCores
The number of cores to run on

keepFiles
TRUE or FALSE, whether you want to keep the bowtie output files

options
Extra options that feed into bowtie2 "options"

Value

A list of aligned and unaligned counts dataframes, respectively

Examples


x <- bt2Align(orfLib = fastaFileName,numCores = 6)

[Package MIPSAlign version 0.0.0.9000 Index]