filterSeqs {MIPSAlign}	R Documentation

Filter Fastq files

Description

Performs a variety of filtering steps to clean Fastq files

Usage


filterSeqs(
  fileDir = getwd(),
  leftPattern = "",
  rightPattern = "",
  minLen = 20,
  trimTo = 40,
  numCores = 1
)

Arguments

fileDir
The file directory with Fastq files, default is current working directory

leftPattern
The 3' adapter to trim sequences by

rightPattern
The 5' adapter to trim sequences by

minLen
All sequences below this minimum length requirement are filtered out

trimTo
Trims all reads to desired length (determined by the length of the longest sequence)

numCores
The number of cores to run on

Examples


filterSeqs(trimTo = 30)

[Package MIPSAlign version 0.0.0.9000 Index]