matchAlign {MIPSAlign}	R Documentation

Perfect match aligner

Description

Generates counts on perfect sequence matches

Usage


matchAlign(seqData = list.files(pattern = "*.fastq.gz$"), orfLib, numCores = 1)

Arguments

seqData
The list of Fastq files you want to align

orfLib
The reference Fasta alignment file

numCores
The number of cores to run on

Value

A list of aligned and unaligned counts dataframes, respectively

Examples


x <- matchAlign(orfLib = fastaFileName,numCores = 6)

[Package MIPSAlign version 0.0.0.9000 Index]